Fix 'hl-printf-regexp' to allow field widths with non-trailing zeros
A trailing zero is part of the "Flags" field, so the "Minimum field
width" field must not start with zero; however, it can contain zeros as
long as they are not the first character.

Before this commit, both "%-10s" and "%-05s" were invalid; after this
commit, "%-10s" would be valid while "%-05s" would still be invalid.